BugsLifeNo23 修正完了
BugsLifeNo23 修正完了

+++ src/main/java/com/example/model/Product.java
@@ -1,7 +1,6 @@
 package com.example.model;
 
 import java.io.Serializable;
-import java.lang.String;
 import java.util.List;
 
 import com.example.form.ProductForm;
@@ -15,10 +14,10 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
-import jakarta.persistence.OneToMany;
 
 @Getter
 @Setter
@@ -48,7 +47,7 @@ public class Product extends TimeEntity implements Serializable {
 	private Integer height;
 
 	@Column(name = "price", nullable = false)
-	private Integer price;
+	private Double price;
 
 	@Column(name = "tax_type", nullable = false)
 	private Integer taxType;
@@ -64,7 +63,7 @@ public class Product extends TimeEntity implements Serializable {
 		this.setCode(form.getCode());
 		this.setWeight(form.getWeight());
 		this.setHeight(form.getHeight());
-		this.setPrice(form.getPrice());
+		this.setPrice(form.getPrice().doubleValue());
 		this.setTaxType(form.getTaxType());
 	}
